server/requirements.txt
mcp-server>=0.1.4
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
openai>=1.3.0
